storageversionmigrator: move regression test into existing test file
Address review feedback: consolidate TestSyncDoesNotPanicWhenResourceNotInMapper
into storageversionmigrator_test.go instead of a separate file.
